moved database code from "server" package to "database" package
