+++ cmd/server/run.go
@@ -14,14 +14,14 @@ import (
 	"github.com/ohsu-comp-bio/funnel/compute/scheduler"
 	"github.com/ohsu-comp-bio/funnel/compute/slurm"
 	"github.com/ohsu-comp-bio/funnel/config"
+	"github.com/ohsu-comp-bio/funnel/database/boltdb"
+	"github.com/ohsu-comp-bio/funnel/database/datastore"
+	"github.com/ohsu-comp-bio/funnel/database/dynamodb"
+	"github.com/ohsu-comp-bio/funnel/database/elastic"
+	"github.com/ohsu-comp-bio/funnel/database/mongodb"
 	"github.com/ohsu-comp-bio/funnel/events"
 	"github.com/ohsu-comp-bio/funnel/logger"
 	"github.com/ohsu-comp-bio/funnel/server"
-	"github.com/ohsu-comp-bio/funnel/server/boltdb"
-	"github.com/ohsu-comp-bio/funnel/server/datastore"
-	"github.com/ohsu-comp-bio/funnel/server/dynamodb"
-	"github.com/ohsu-comp-bio/funnel/server/elastic"
-	"github.com/ohsu-comp-bio/funnel/server/mongodb"
 	"github.com/ohsu-comp-bio/funnel/tes"
 )
