+++ cmd/worker/run.go
@@ -6,12 +6,12 @@ import (
 	"strings"
 
 	"github.com/ohsu-comp-bio/funnel/config"
+	"github.com/ohsu-comp-bio/funnel/database/datastore"
+	"github.com/ohsu-comp-bio/funnel/database/dynamodb"
+	"github.com/ohsu-comp-bio/funnel/database/elastic"
+	"github.com/ohsu-comp-bio/funnel/database/mongodb"
 	"github.com/ohsu-comp-bio/funnel/events"
 	"github.com/ohsu-comp-bio/funnel/logger"
-	"github.com/ohsu-comp-bio/funnel/server/datastore"
-	"github.com/ohsu-comp-bio/funnel/server/dynamodb"
-	"github.com/ohsu-comp-bio/funnel/server/elastic"
-	"github.com/ohsu-comp-bio/funnel/server/mongodb"
 	"github.com/ohsu-comp-bio/funnel/storage"
 	"github.com/ohsu-comp-bio/funnel/tes"
 	"github.com/ohsu-comp-bio/funnel/worker"
