+++ database/dynamodb/util.go
